Move and update files from experiment to new
