+++ src/app/new/ChatGPTButton.tsx
